python_server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
websockets==12.0
autogen-agentchat>=0.2.0
autogen-ext[openai]>=0.2.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.0
aiofiles==23.2.0
